- add cloud service

+++ controllers/productController.js
@@ -1,8 +1,10 @@
 const Product = require("../models/Product");
-const upload = require('../config/multer');
+// const upload = require('../config/multer');
 const fs = require('fs');
 const path = require('path');
-
+const cloudinary = require('cloudinary').v2;
+const multer = require('multer');
+const { CloudinaryStorage } = require('multer-storage-cloudinary');
 // Utility function to validate ObjectId
 const isValidObjectId = (id) => /^[0-9a-fA-F]{24}$/.test(id);
 
@@ -74,16 +76,37 @@ const handleTotalFileSize = (req, res, next) => {
   next();
 };
 
+
+// // Configure Cloudinary
+cloudinary.config({
+  cloud_name: 'dzau808tk',
+  api_key: '475953758675764',
+  api_secret: 'Sdq1SfomO9B0t283aF7QlZPO3i4' // Click 'View API Keys' above to copy your API secret
+});
+
+const storage = new CloudinaryStorage({
+  cloudinary: cloudinary,
+  params: {
+    folder: 'uploads', // The name of the folder in your Cloudinary account
+    allowed_formats: ['jpg', 'png', 'jpeg'],
+  },
+});
+
+const upload = multer({ storage: storage });
+
+
 // Upload images
-exports.uploadImages = [
-  upload.array('images'),
-  handleFileValidationError,
-  handleTotalFileSize,
-  (req, res) => {
-    console.log('Uploaded files:', req.files);
-    res.status(200).json({ message: 'Files uploaded successfully', files: req.files });
-  }
-];
+
+exports.uploadImages = async (req, res) => {
+    upload.array('images', 5)(req, res, (err) => {
+        if (err instanceof multer.MulterError) {
+        return res.status(400).json({ message: err.message });
+        } else if (err) {
+        return res.status(500).json({ message: err.message });
+        }
+        res.status(200).json({ message: 'Images uploaded successfully', files: req.files });
+    });
+}
 
 // Delete a file
 exports.deleteFile = (req, res) => {
@@ -146,6 +169,14 @@ exports.deleteProduct = async (req, res) => {
   }
 
   try {
+    // delete product images from cloudinary
+    const product = await Product.findById(productId);
+    if (!product) {
+      return res.status(404).json({ message: 'Product not found.' });
+    }
+    const imageIds = product.image.map((img) => img.filename);
+    await cloudinary.api.delete_resources(imageIds);
+
     const result = await Product.findByIdAndDelete(productId);
     if (!result) {
       return res.status(404).json({ message: 'Product not found.' });
